oooast-tools-50 Remove duplication code in production line executor - fixed
